Fixed -- PR comments
