+++ tests/drmcp/unit/dynamic_prompts/test_register.py
@@ -14,57 +14,14 @@
 
 """Tests for external prompt registration."""
 
-from collections.abc import Iterator
-from unittest.mock import patch
-
 import pytest
 
-from datarobot_genai.drmcp.core.dynamic_prompts.dr_lib import DrPrompt
-from datarobot_genai.drmcp.core.dynamic_prompts.dr_lib import DrPromptVersion
 from datarobot_genai.drmcp.core.dynamic_prompts.dr_lib import DrVariable
 from datarobot_genai.drmcp.core.dynamic_prompts.register import make_prompt_function
 from datarobot_genai.drmcp.core.dynamic_prompts.register import (
     register_prompts_from_datarobot_prompt_management,
 )
 from datarobot_genai.drmcp.core.mcp_instance import mcp
-
-
-@pytest.fixture
-def prompt_template_id_ok() -> str:
-    return "69086ea4834952718366b2ce"
-
-
-@pytest.fixture
-def prompt_template_version_id_ok() -> str:
-    return "69086ea4b65d70489c5b198d"
-
-
-@pytest.fixture
-def get_prompt_template_mock(
-    prompt_template_id_ok: str, prompt_template_version_id_ok: str
-) -> Iterator[None]:
-    """Set up all API endpoint mocks."""
-    dr_prompt_version = DrPromptVersion(
-        id=prompt_template_version_id_ok,
-        version=3,
-        prompt_text="Write greeting for {{name}} in max {{sentences}} sentences.",
-        variables=[
-            DrVariable(name="name", description="Person name"),
-            DrVariable(name="sentences", description="Number of sentences"),
-        ],
-    )
-    dr_prompt = DrPrompt(
-        id=prompt_template_id_ok,
-        name="Dummy prompt name",
-        description="Dummy description",
-    )
-    dr_prompt.get_latest_version = lambda: dr_prompt_version
-
-    with patch(
-        "datarobot_genai.drmcp.core.dynamic_prompts.register.get_datarobot_prompt_templates",
-        return_value=[dr_prompt],
-    ):
-        yield
 
 
 class TestMakePrompt:
